requirements.txt
# Firco XGBoost API Requirements
# High-performance compliance prediction with hierarchical training

# Core ML Libraries
xgboost>=2.0.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0

# Feature Engineering & Data Processing
scipy>=1.11.0
joblib>=1.3.0

# Text Processing
nltk>=3.8.0

# FastAPI and Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic-settings>=2.0.0

# HTTP Client for Testing
requests>=2.31.0
httpx>=0.25.0

# Data Validation & Processing
openpyxl>=3.1.0
xlrd>=2.0.1

# Visualization (optional)
matplotlib>=3.7.0
seaborn>=0.12.0

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Logging and Monitoring
python-json-logger>=2.0.7

# Performance Monitoring
psutil>=5.9.0

# Date/Time Processing
python-dateutil>=2.8.0

# Environment Management
python-dotenv>=1.0.0

# MongoDB Integration
pymongo>=4.5.0
motor>=3.3.0  # Async MongoDB driver for production

# DNS resolver for MongoDB connections
dnspython>=2.4.0

# Optional: GPU support for XGBoost (uncomment if needed)
# xgboost[gpu]>=2.0.0

# Optional: Advanced text processing (uncomment if needed)
# spacy>=3.7.0
# transformers>=4.30.0 